+++ API/Services/Interfaces/IMatchScoresService.cs
@@ -4,8 +4,8 @@ namespace API.Services.Interfaces;
 
 public interface IMatchScoresService : IService<MatchScore>
 {
-	Task<IEnumerable<MatchScore>> GetForGameAsync(int gameId);
-	Task<IEnumerable<MatchScore>> GetForPlayerAsync(int playerId);
+	Task<IEnumerable<MatchScore>> GetForGameAsync(long gameId);
+	Task<IEnumerable<MatchScore>> GetForPlayerAsync(long playerId);
 	Task<int> BulkInsertAsync(IEnumerable<MatchScore> matchScores);
 	Task<int?> CreateIfNotExistsAsync(MatchScore dbScore);
 }
